refactor: rename reward-claim file
