+++ the-graph/subgraphs/staking/src/functions/index.ts
@@ -3,4 +3,4 @@ export * from './staking'
 export * from './subscription'
 export * from './token'
 export * from './user'
-export * from './rewardClaim'
+export * from './reward-claim'
